[FEAT] : 칵테일 별점높은순 조회기능 구현

+++ src/main/java/com/umc/goldenratio/api/dto/request/WriteReviewDto.java
@@ -2,13 +2,15 @@ package com.umc.goldenratio.api.dto.request;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class WriteReviewDto {
     private String content;
-    private int score;
+    private BigDecimal score;
 
 
 }
